Developing a real, standard quote page.
Colors need some work. It would be nice to have 'em transparent. The image of
the thing we're quoting often shows up too far down, so the quote covers it a
lot. Perhaps we'll just have to use those pages judiciously, and only for
shorter quotes. Since the problem with quotes currently is empty space, which
goes away with a long quote anyway, there's justification for that idea.